Pull request #3485: Fix config logger
Merge in SNORT/snort3 from ~VHORBAN/snort3:fix_config_logger to master

Squashed commit of the following:

commit 4ce90eea0b6b7c75f4321c3cabdc6781178291a9
Author: Oleksandr Serhiienko <oserhiie@cisco.com>
Date:   Wed Jun 15 17:21:26 2022 +0300

    build: remove unnecessary type casts

commit 6cda44321578d31de30524a5b8a50ce7713ecea9
Author: Oleksandr Serhiienko <oserhiie@cisco.com>
Date:   Wed Jun 15 17:02:52 2022 +0300

    log: add log_value and log_limit overloads with built-in integer types

        Using built-in integer types in overloads of ConfigLogger::log_value
        and ConfigLogger::log_limit resolves possible ambiguity over different
        platforms in case of platform-dependent integer types like size_t

+++ src/network_inspectors/appid/appid_config.cc
@@ -87,7 +87,7 @@ void AppIdConfig::show() const
 
     ConfigLogger::log_flag("log_all_sessions", log_all_sessions);
     ConfigLogger::log_flag("log_stats", log_stats);
-    ConfigLogger::log_value("memcap", static_cast<uint64_t>(memcap));
+    ConfigLogger::log_value("memcap", memcap);
 }
 
 void AppIdContext::pterm()
