+++ src/service_inspectors/netflow/netflow.cc
@@ -829,8 +829,8 @@ static void show_device(const NetFlowRule& d, bool is_exclude)
 
 void NetFlowInspector::show(const SnortConfig*) const
 {
-    ConfigLogger::log_value("flow_memcap", (uint64_t)config->flow_memcap);
-    ConfigLogger::log_value("template_memcap", (uint64_t)config->template_memcap);
+    ConfigLogger::log_value("flow_memcap", config->flow_memcap);
+    ConfigLogger::log_value("template_memcap", config->template_memcap);
     ConfigLogger::log_value("dump_file", config->dump_file);
     ConfigLogger::log_value("update_timeout", config->update_timeout);
     bool log_header = true;
